requirements.txt
psycopg2-binary
requests
pandas
numpy
xgboost
scikit-learn
